add -Wall into compiler flags. fixed all warnings. fixed _deleteCandidate when tails are involved.

+++ src/ime-core/imi_view_classic.h
@@ -62,7 +62,7 @@ public:
     virtual int  onCandidateSelectRequest(int index);
 
 private:
-    static int top_candidate_threshold;
+    static size_t top_candidate_threshold;
 
     unsigned m_cursorFrIdx;
     unsigned m_candiFrIdx;
